problem with the profile main, may white space sa baba. + improved post design
