+++ lib/screens/ProfileScreen/profile_main.dart
@@ -72,14 +72,14 @@ class _ContentState extends State<Content> {
           showGeneralDialog(
             context: context,
             pageBuilder: (BuildContext context, Animation<double> animation,
-                    Animation<double> secondaryAnimation) =>
+                Animation<double> secondaryAnimation) =>
                 SizedBox(
-              width: MediaQuery.of(context).size.width,
-              child: const CreatePostScreen(),
-            ),
+                  width: MediaQuery.of(context).size.width,
+                  child: const CreatePostScreen(),
+                ),
             barrierDismissible: true,
             barrierLabel:
-                MaterialLocalizations.of(context).modalBarrierDismissLabel,
+            MaterialLocalizations.of(context).modalBarrierDismissLabel,
             barrierColor: Colors.black45,
             transitionDuration: const Duration(milliseconds: 250),
             transitionBuilder: (context, animation, secondaryAnimation, child) {
@@ -127,14 +127,14 @@ class _ContentState extends State<Content> {
               showGeneralDialog(
                 context: context,
                 pageBuilder: (BuildContext context, Animation<double> animation,
-                        Animation<double> secondaryAnimation) =>
+                    Animation<double> secondaryAnimation) =>
                     Container(
-                  width: MediaQuery.of(context).size.width,
-                  child: profileEdit(),
-                ),
+                      width: MediaQuery.of(context).size.width,
+                      child: profileEdit(),
+                    ),
                 barrierDismissible: true,
                 barrierLabel:
-                    MaterialLocalizations.of(context).modalBarrierDismissLabel,
+                MaterialLocalizations.of(context).modalBarrierDismissLabel,
                 barrierColor: Colors.black45,
                 transitionDuration: const Duration(milliseconds: 250),
                 transitionBuilder:
@@ -151,12 +151,12 @@ class _ContentState extends State<Content> {
             },
             style: ButtonStyle(
               backgroundColor: MaterialStateProperty.resolveWith<Color>(
-                (Set<MaterialState> states) {
+                    (Set<MaterialState> states) {
                   return Colors.white;
                 },
               ),
               side: MaterialStateProperty.resolveWith<BorderSide>(
-                (Set<MaterialState> states) {
+                    (Set<MaterialState> states) {
                   return const BorderSide(
                     color: Color(0xFF1E7251),
                     width: 2.0,
@@ -182,195 +182,198 @@ class _ContentState extends State<Content> {
   }
 
   Widget buildCoverImage() => Container(
-        child: (widget.userData?.header != null &&
-                widget.userData!.header.isNotEmpty)
-            ? Image.memory(
-                base64Decode(widget.userData!.header),
-                width: double.infinity,
-                height: coverHeight,
-                fit: BoxFit.cover,
-              )
-            : Container(
-                width: double.infinity,
-                height: coverHeight,
-                color: Color(0xFF1E7251),
-              ),
-      );
+    child: (widget.userData?.header != null &&
+        widget.userData!.header.isNotEmpty)
+        ? Image.memory(
+      base64Decode(widget.userData!.header),
+      width: double.infinity,
+      height: coverHeight,
+      fit: BoxFit.cover,
+    )
+        : Container(
+      width: double.infinity,
+      height: coverHeight,
+      color: Color(0xFF1E7251),
+    ),
+  );
 
   Widget buildProfileImage() => Container(
-        padding: const EdgeInsets.all(5),
-        decoration: const BoxDecoration(
+    padding: const EdgeInsets.all(5),
+    decoration: const BoxDecoration(
+      color: Colors.white,
+      shape: BoxShape.circle,
+    ),
+    child: CircleAvatar(
+      radius: profileHeight / 2,
+      backgroundColor: Color(0xFF1E7251),
+      backgroundImage: widget.userData?.icon != null
+          ? MemoryImage(base64Decode(widget.userData!.icon))
+          : null,
+      child: (widget.userData?.icon != null && widget.userData?.icon == '')
+          ? Text(
+        widget.userData?.fname[0].toUpperCase() ?? '',
+        style: TextStyle(
+          fontSize: 40.0,
           color: Colors.white,
-          shape: BoxShape.circle,
         ),
-        child: CircleAvatar(
-          radius: profileHeight / 2,
-          backgroundColor: Color(0xFF1E7251),
-          backgroundImage: widget.userData?.icon != null
-              ? MemoryImage(base64Decode(widget.userData!.icon))
-              : null,
-          child: (widget.userData?.icon != null && widget.userData?.icon == '')
-              ? Text(
-                  widget.userData?.fname[0].toUpperCase() ?? '',
-                  style: TextStyle(
-                    fontSize: 40.0,
-                    color: Colors.white,
-                  ),
-                )
-              : null,
-        ),
-      );
+      )
+          : null,
+    ),
+  );
 
   Widget buildProfileInfo() => Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
+    crossAxisAlignment: CrossAxisAlignment.start,
+    children: [
+      const SizedBox(
+        height: 8,
+      ),
+      Text(
+        '${widget.userData!.fname} ${widget.userData!.lname}',
+        style: TextStyle(
+          fontSize: 28,
+          fontWeight: FontWeight.bold,
+        ),
+      ),
+      Text(
+        '@${widget.userData?.username}',
+        style: TextStyle(
+          fontSize: 18,
+        ),
+      ),
+      Text(
+        widget.userData!.bio,
+        style: TextStyle(
+          fontSize: 18,
+        ),
+      ),
+      Row(
         children: [
-          const SizedBox(
-            height: 8,
+          Icon(
+            Icons.location_on,
+            size: 20,
+            color: Color(0xFF2C2C30),
+          ),
+          SizedBox(
+            width: 5,
           ),
           Text(
-            '${widget.userData!.fname} ${widget.userData!.lname}',
-            style: TextStyle(
-              fontSize: 28,
-              fontWeight: FontWeight.bold,
-            ),
-          ),
-          Text(
-            '@${widget.userData?.username}',
+            widget.userData!.location,
             style: TextStyle(
               fontSize: 18,
             ),
-          ),
-          Text(
-            widget.userData!.bio,
-            style: TextStyle(
-              fontSize: 18,
-            ),
-          ),
-          Row(
-            children: [
-              Icon(
-                Icons.location_on,
-                size: 20,
-                color: Color(0xFF2C2C30),
-              ),
-              SizedBox(
-                width: 5,
-              ),
-              Text(
-                widget.userData!.location,
-                style: TextStyle(
-                  fontSize: 18,
-                ),
-              ),
-            ],
-          ),
-          Row(
-            children: [
-              TextButton(
-                  onPressed: () {},
-                  style: ButtonStyle(
-                    padding: MaterialStateProperty.all(EdgeInsets.zero),
-                    overlayColor: MaterialStateProperty.resolveWith<Color>(
-                      (Set<MaterialState> states) {
-                        if (states.contains(MaterialState.pressed)) {
-                          return Colors.transparent;
-                        }
-                        return Colors.transparent;
-                      },
-                    ),
-                    foregroundColor: MaterialStateProperty.resolveWith<Color>(
-                      (Set<MaterialState> states) {
-                        if (states.contains(MaterialState.pressed)) {
-                          return const Color(0xFF4E4C4C).withOpacity(0.5);
-                        }
-                        return const Color(0xFF4E4C4C);
-                      },
-                    ),
-                  ),
-                  child: const Row(
-                    children: [
-                      Text(
-                        '69',
-                        style: TextStyle(
-                          fontSize: 18,
-                          fontWeight: FontWeight.bold,
-                        ),
-                      ),
-                      SizedBox(
-                        width: 4,
-                      ),
-                      Text(
-                        'Friends',
-                        style: TextStyle(
-                          fontSize: 18,
-                          fontWeight: FontWeight.bold,
-                        ),
-                      ),
-                    ],
-                  )),
-              const SizedBox(
-                width: 10,
-              ),
-              TextButton(
-                  onPressed: () {},
-                  style: ButtonStyle(
-                    padding: MaterialStateProperty.all(EdgeInsets.zero),
-                    overlayColor: MaterialStateProperty.resolveWith<Color>(
-                      (Set<MaterialState> states) {
-                        if (states.contains(MaterialState.pressed)) {
-                          return Colors.transparent;
-                        }
-                        return Colors.transparent;
-                      },
-                    ),
-                    foregroundColor: MaterialStateProperty.resolveWith<Color>(
-                      (Set<MaterialState> states) {
-                        if (states.contains(MaterialState.pressed)) {
-                          return const Color(0xFF4E4C4C).withOpacity(0.5);
-                        }
-                        return const Color(0xFF4E4C4C);
-                      },
-                    ),
-                  ),
-                  child: const Row(
-                    children: [
-                      Text(
-                        '420',
-                        style: TextStyle(
-                          fontSize: 18,
-                          fontWeight: FontWeight.bold,
-                        ),
-                      ),
-                      SizedBox(
-                        width: 4,
-                      ),
-                      Text(
-                        'Following',
-                        style: TextStyle(
-                          fontSize: 18,
-                          fontWeight: FontWeight.bold,
-                        ),
-                      ),
-                    ],
-                  )),
-            ],
           ),
         ],
-      );
+      ),
+      Row(
+        children: [
+          TextButton(
+              onPressed: () {},
+              style: ButtonStyle(
+                padding: MaterialStateProperty.all(EdgeInsets.zero),
+                overlayColor: MaterialStateProperty.resolveWith<Color>(
+                      (Set<MaterialState> states) {
+                    if (states.contains(MaterialState.pressed)) {
+                      return Colors.transparent;
+                    }
+                    return Colors.transparent;
+                  },
+                ),
+                foregroundColor: MaterialStateProperty.resolveWith<Color>(
+                      (Set<MaterialState> states) {
+                    if (states.contains(MaterialState.pressed)) {
+                      return const Color(0xFF4E4C4C).withOpacity(0.5);
+                    }
+                    return const Color(0xFF4E4C4C);
+                  },
+                ),
+              ),
+              child: const Row(
+                children: [
+                  Text(
+                    '69',
+                    style: TextStyle(
+                      fontSize: 18,
+                      fontWeight: FontWeight.bold,
+                    ),
+                  ),
+                  SizedBox(
+                    width: 4,
+                  ),
+                  Text(
+                    'Friends',
+                    style: TextStyle(
+                      fontSize: 18,
+                      fontWeight: FontWeight.bold,
+                    ),
+                  ),
+                ],
+              )),
+          const SizedBox(
+            width: 10,
+          ),
+          TextButton(
+              onPressed: () {},
+              style: ButtonStyle(
+                padding: MaterialStateProperty.all(EdgeInsets.zero),
+                overlayColor: MaterialStateProperty.resolveWith<Color>(
+                      (Set<MaterialState> states) {
+                    if (states.contains(MaterialState.pressed)) {
+                      return Colors.transparent;
+                    }
+                    return Colors.transparent;
+                  },
+                ),
+                foregroundColor: MaterialStateProperty.resolveWith<Color>(
+                      (Set<MaterialState> states) {
+                    if (states.contains(MaterialState.pressed)) {
+                      return const Color(0xFF4E4C4C).withOpacity(0.5);
+                    }
+                    return const Color(0xFF4E4C4C);
+                  },
+                ),
+              ),
+              child: const Row(
+                children: [
+                  Text(
+                    '420',
+                    style: TextStyle(
+                      fontSize: 18,
+                      fontWeight: FontWeight.bold,
+                    ),
+                  ),
+                  SizedBox(
+                    width: 4,
+                  ),
+                  Text(
+                    'Following',
+                    style: TextStyle(
+                      fontSize: 18,
+                      fontWeight: FontWeight.bold,
+                    ),
+                  ),
+                ],
+              )),
+        ],
+      ),
+    ],
+  );
 
   Widget buildContent() {
     return Container(
       height: MediaQuery.of(context).size.height,
       child: Column(
         children: [
-          Row(
-            mainAxisAlignment: MainAxisAlignment.center,
-            children: [
-              buildContentButton('Posts', 0),
-              buildContentButton('Replies', 1),
-              buildContentButton("GTHR'D", 2),
-              buildContentButton('Likes', 3),
-            ],
+          Padding(
+            padding: const EdgeInsets.only(bottom: 30),
+            child: Row(
+              mainAxisAlignment: MainAxisAlignment.spaceEvenly,
+              children: [
+                buildContentButton('Posts', 0),
+                buildContentButton('Replies', 1),
+                buildContentButton("GTHR'D", 2),
+                buildContentButton('Likes', 3),
+              ],
+            ),
           ),
           if (selectedIndex == 0) buildPostsContent(),
           if (selectedIndex == 1) buildRepliesContent(),
@@ -388,9 +391,9 @@ class _ContentState extends State<Content> {
       decoration: BoxDecoration(
           border: Border(
               bottom: BorderSide(
-        color: isActive ? const Color(0xFFFF4E1A) : Colors.transparent,
-        width: 5.0,
-      ))),
+                color: isActive ? const Color(0xFFFF4E1A) : Colors.transparent,
+                width: 5.0,
+              ))),
       child: TextButton(
         onPressed: () {
           setState(() {
@@ -399,7 +402,7 @@ class _ContentState extends State<Content> {
         },
         style: ButtonStyle(
           overlayColor: MaterialStateProperty.resolveWith<Color>(
-            (Set<MaterialState> states) {
+                (Set<MaterialState> states) {
               if (states.contains(MaterialState.pressed)) {
                 return Colors.transparent;
               }
@@ -407,7 +410,7 @@ class _ContentState extends State<Content> {
             },
           ),
           foregroundColor: MaterialStateProperty.resolveWith<Color>(
-            (Set<MaterialState> states) {
+                (Set<MaterialState> states) {
               if (states.contains(MaterialState.pressed)) {
                 return const Color(0xFF4E4C4C).withOpacity(0.5);
               }
@@ -429,34 +432,134 @@ class _ContentState extends State<Content> {
       builder: (context, snapshot) {
         if (snapshot.hasData) {
           List<Post> posts = snapshot.data!;
+          posts.sort((a, b) => b.timestamp.compareTo(a.timestamp));
+          if (posts.isEmpty) {
+            return const Padding(
+              padding: EdgeInsets.symmetric(vertical: 20),
+              child: Column(
+                children: [
+                  Text('Nothing to see here, fam. Why not post something?')
+                ],
+              ),
+            );
+          } else {
+            return Expanded(
+              child: ListView.builder(
+                  physics: const NeverScrollableScrollPhysics(),
+                  itemCount: posts.length,
+                  itemBuilder: (context, index) {
+                    return Row(
+                      crossAxisAlignment: CrossAxisAlignment.start,
+                      children: [
+                        CircleAvatar(
+                          radius: 20,
+                          backgroundColor: Color(0xFF1E7251),
+                          backgroundImage: widget.userData?.icon != null
+                              ? MemoryImage(base64Decode(widget.userData!.icon))
+                              : null,
+                          child: (widget.userData?.icon != null &&
+                              widget.userData?.icon == '')
+                              ? Text(
+                            widget.userData?.fname[0].toUpperCase() ?? '',
+                            style: TextStyle(
+                              fontSize: 40.0,
+                              color: Colors.white,
+                            ),
+                          )
+                              : null,
+                        ),
+                        Flexible(
+                          child: Column(
+                            crossAxisAlignment: CrossAxisAlignment.start,
+                            children: [
+                              Row(
+                                mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                                children: [
+                                  Text(
+                                    '@${widget.userData!.username} · ${timeago.format(posts[index].timestamp, locale: 'en_short')}',
+                                    style: TextStyle(
+                                      color: Colors.grey,
+                                      fontSize: 16.0,
+                                    ),
+                                  ),
+                                  IconButton(
+                                      onPressed: (){},
+                                      icon: Icon(Icons.more_horiz)
+                                  )
+                                ],
+                              ),
+                              Text(
+                                posts[index].content,
+                                style: TextStyle(
+                                  fontSize: 18.0,
+                                ),
+                              ),
+                            ],
+                          ),
+                        ),
+                      ],
+                    );
+                  }
+              ),
+            );
+          }
           return Expanded(
             child: ListView.builder(
-              itemCount: posts.length,
-              itemBuilder: (context, index) {
-                return ListTile(
-                  leading: CircleAvatar(
-                      backgroundImage: widget.userData?.icon != null
-                          ? MemoryImage(base64Decode(widget.userData!.icon))
-                          : null,
-                  ),
-                  title: Text(
-                    posts[index].content,
-                    style: TextStyle(
-                      color: Colors.blue,
-                      fontSize: 20.0,
-                      fontWeight: FontWeight.bold,
-                    ),
-                  ),
-                  subtitle: Text(
-                    timeago.format(posts[index].timestamp),
-                    style: TextStyle(
-                      color: Colors.grey,
-                      fontSize: 16.0,
-                    ),
-                  ),
-                  trailing: Icon(Icons.more_horiz),
-                );
-              },
+                physics: const NeverScrollableScrollPhysics(),
+                itemCount: posts.length,
+                itemBuilder: (context, index) {
+                  return Row(
+                    crossAxisAlignment: CrossAxisAlignment.start,
+                    children: [
+                      CircleAvatar(
+                        radius: 20,
+                        backgroundColor: Color(0xFF1E7251),
+                        backgroundImage: widget.userData?.icon != null
+                            ? MemoryImage(base64Decode(widget.userData!.icon))
+                            : null,
+                        child: (widget.userData?.icon != null &&
+                            widget.userData?.icon == '')
+                            ? Text(
+                          widget.userData?.fname[0].toUpperCase() ?? '',
+                          style: TextStyle(
+                            fontSize: 40.0,
+                            color: Colors.white,
+                          ),
+                        )
+                            : null,
+                      ),
+                      Flexible(
+                        child: Column(
+                          crossAxisAlignment: CrossAxisAlignment.start,
+                          children: [
+                            Row(
+                              mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                              children: [
+                                Text(
+                                  '@${widget.userData!.username} · ${timeago.format(posts[index].timestamp, locale: 'en_short')}',
+                                  style: TextStyle(
+                                    color: Colors.grey,
+                                    fontSize: 16.0,
+                                  ),
+                                ),
+                                IconButton(
+                                    onPressed: (){},
+                                    icon: Icon(Icons.more_horiz)
+                                )
+                              ],
+                            ),
+                            Text(
+                              posts[index].content,
+                              style: TextStyle(
+                                fontSize: 18.0,
+                              ),
+                            ),
+                          ],
+                        ),
+                      ),
+                    ],
+                  );
+                }
             ),
           );
         } else if (snapshot.hasError) {
@@ -469,30 +572,30 @@ class _ContentState extends State<Content> {
   }
 
   Widget buildRepliesContent() => const Padding(
-        padding: EdgeInsets.symmetric(vertical: 20),
-        child: Column(
-          children: [
-            Text('Nothing to see here, fam. Why not reply to someone?')
-          ],
-        ),
-      );
+    padding: EdgeInsets.symmetric(vertical: 20),
+    child: Column(
+      children: [
+        Text('Nothing to see here, fam. Why not reply to someone?')
+      ],
+    ),
+  );
 
   Widget buildGthrContent() => Padding(
-        padding: const EdgeInsets.symmetric(vertical: 20),
-        child: Column(
-          children: [
-            Image.network(
-                'https://asset-ent.abs-cbn.com/ent/entertainment/media/onemusic/lovebox.jpg?ext=.jpg'),
-            Image.network(
-                'https://scontent.fmnl8-1.fna.fbcdn.net/v/t39.30808-6/401836757_737590591748457_5740288807220724461_n.jpg?stp=dst-jpg_p843x403&_nc_cat=106&ccb=1-7&_nc_sid=5f2048&_nc_ohc=zW3lI9lyNNkAb7qpwLC&_nc_ht=scontent.fmnl8-1.fna&cb_e2o_trans=q&oh=00_AfARb92ZkIqxiMZf_HvGqwaLlD0l7sqe-SUmUguq4gSjNQ&oe=66259EC4'),
-          ],
-        ),
-      );
+    padding: const EdgeInsets.symmetric(vertical: 20),
+    child: Column(
+      children: [
+        Image.network(
+            'https://asset-ent.abs-cbn.com/ent/entertainment/media/onemusic/lovebox.jpg?ext=.jpg'),
+        Image.network(
+            'https://scontent.fmnl8-1.fna.fbcdn.net/v/t39.30808-6/401836757_737590591748457_5740288807220724461_n.jpg?stp=dst-jpg_p843x403&_nc_cat=106&ccb=1-7&_nc_sid=5f2048&_nc_ohc=zW3lI9lyNNkAb7qpwLC&_nc_ht=scontent.fmnl8-1.fna&cb_e2o_trans=q&oh=00_AfARb92ZkIqxiMZf_HvGqwaLlD0l7sqe-SUmUguq4gSjNQ&oe=66259EC4'),
+      ],
+    ),
+  );
 
   Widget buildLikesContent() => const Padding(
-        padding: EdgeInsets.symmetric(vertical: 20),
-        child: Column(
-          children: [Text('Nothing to see here, fam. Why not like something?')],
-        ),
-      );
+    padding: EdgeInsets.symmetric(vertical: 20),
+    child: Column(
+      children: [Text('Nothing to see here, fam. Why not like something?')],
+    ),
+  );
 }
