+++ lib/screens/authenticate/SignUpScreen/signup1.dart
@@ -1,12 +1,11 @@
 import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 import 'package:gthr/services/auth.dart';
-
-import '../../../shared/loading.dart';
+import 'package:gthr/shared/loading.dart';
 
 class SignUpScreen extends StatefulWidget {
-
   final void Function(bool, bool) toggleView;
+
   const SignUpScreen({super.key, required this.toggleView});
 
   @override
@@ -14,7 +13,6 @@ class SignUpScreen extends StatefulWidget {
 }
 
 class _SignUpScreenState extends State<SignUpScreen> {
-
   final AuthService _auth = AuthService();
   final GlobalKey<FormState> _formKey = GlobalKey<FormState>();
   bool loading = false;
@@ -41,79 +39,24 @@ class _SignUpScreenState extends State<SignUpScreen> {
           child: Column(
             children: <Widget>[
               const SizedBox(height: 10),
-              TextFormField(
-                validator: (val) => val!.isEmpty ? "Enter your First Name" : null,
-                onChanged: (val){
-                  setState(() => fname = val);
-                },
-                decoration: const InputDecoration(
-                  labelText: 'First Name',
-                  border: OutlineInputBorder(),
-                ),
-              ),
+              buildTextField('First Name', 'Enter your First Name', validator: (val) => val!.isEmpty ? "Enter your First Name" : null),
               const SizedBox(height: 15),
-              TextFormField(
-                validator: (val) => val!.isEmpty ? "Enter your Last Name" : null,
-                onChanged: (val){
-                  setState(() => lname = val);
-                },
-                decoration: const InputDecoration(
-                  labelText: 'Last Name',
-                  border: OutlineInputBorder(),
-                ),
-              ),
+              buildTextField('Last Name', 'Enter your Last Name', validator: (val) => val!.isEmpty ? "Enter your Last Name" : null),
               const SizedBox(height: 15),
-              TextFormField(
-                validator: (val) => val!.isEmpty ? "Enter a username" : null,
-                onChanged: (val){
-                  setState(() => username = val);
-                },
-                decoration: const InputDecoration(
-                  labelText: 'Username',
-                  border: OutlineInputBorder(),
-                ),
-              ),
+              buildTextField('Username', 'Enter a username', validator: (val) => val!.isEmpty ? "Enter a username" : null),
               const SizedBox(height: 15),
-              TextFormField(
-                validator: (val) => val!.isEmpty ? "Enter an email" : null,
-                onChanged: (val){
-                  setState(() => email = val);
-                },
-                decoration: const InputDecoration(
-                  labelText: 'E-mail',
-                  border: OutlineInputBorder(),
-                ),
-              ),
+              buildTextField('E-mail', 'Enter an email', validator: (val) => !RegExp(r'\S+@\S+\.\S+').hasMatch(val!) ? "Enter a valid email address" : null),
               const SizedBox(height: 15),
-              TextFormField(
-                validator: (val) => val!.length < 6 ? "Enter a password 6+ chars long" : null,
-                onChanged: (val){
-                  setState(() => password = val);
-                },
-                obscureText: true,
-                decoration: const InputDecoration(
-                  labelText: 'Password',
-                  border: OutlineInputBorder(),
-                ),
-              ),
+              buildTextField('Password', 'Enter a password 6+ chars long', obscureText: true, validator: validatePassword),
               const SizedBox(height: 15),
-              TextFormField(
-                validator: (val) => val!.isEmpty ? "Enter a University" : null,
-                onChanged: (val){
-                  setState(() => uni = val);
-                },
-                decoration: const InputDecoration(
-                  labelText: 'University',
-                  border: OutlineInputBorder(),
-                ),
-              ),
+              buildTextField('University', 'Enter a University', validator: (val) => val!.isEmpty ? "Enter a University" : null),
               const SizedBox(height: 30),
               ElevatedButton(
                 onPressed: () async {
-                  setState(() => loading = true);
-                  if (_formKey.currentState!.validate()){
+                  if (_formKey.currentState!.validate()) {
+                    setState(() => loading = true);
                     dynamic result = await _auth.registerWithEmailAndPassword(fname, lname, username, email, password, uni);
-                    if (result == null){
+                    if (result == null) {
                       setState(() {
                         error = 'Please supply a valid email';
                         loading = false;
@@ -124,11 +67,12 @@ class _SignUpScreenState extends State<SignUpScreen> {
                   }
                 },
                 style: ElevatedButton.styleFrom(
-                  foregroundColor: Colors.white, backgroundColor: const Color(0xff1E7251), // text color
-                  minimumSize: const Size(double.infinity, 50), // button size
+                  foregroundColor: Colors.white, backgroundColor: const Color(0xff1E7251),
+                  minimumSize: const Size(double.infinity, 50),
                 ),
                 child: const Text('Sign Up'),
               ),
+              SizedBox(height: 10),
               RichText(
                 text: TextSpan(
                   style: const TextStyle(color: Colors.black, fontSize: 14.0),
@@ -137,10 +81,9 @@ class _SignUpScreenState extends State<SignUpScreen> {
                     TextSpan(
                       text: 'Log In',
                       style: const TextStyle(color: Color(0xffFB5017), fontWeight: FontWeight.bold, decoration: TextDecoration.underline,),
-                      recognizer: TapGestureRecognizer()
-                        ..onTap = () {
-                          widget.toggleView(true, true);
-                        },
+                      recognizer: TapGestureRecognizer()..onTap = () {
+                        widget.toggleView(true, true);
+                      },
                     ),
                   ],
                 ),
@@ -152,35 +95,47 @@ class _SignUpScreenState extends State<SignUpScreen> {
     );
   }
 
-  Widget _buildStepCircle(String number, Color color) {
-    return CircleAvatar(
-      backgroundColor: color,
-      child: Text(number, style: const TextStyle(color: Colors.white)),
-    );
-  }
-
-  Widget _buildLineConnector(List<Color> colors) {
-    return Expanded(
-      child: Container(
-        height: 5,
-        decoration: BoxDecoration(
-          gradient: LinearGradient(
-            colors: colors,
-            begin: Alignment.centerLeft,
-            end: Alignment.centerRight,
-          ),
-        ),
-      ),
-    );
-  }
-
-  Widget _buildTextField(String label, {bool obscureText = false}) {
-    return TextField(
+  Widget buildTextField(String label, String errorMessage, {bool obscureText = false, String? Function(String?)? validator}) {
+    return TextFormField(
+      validator: validator,
+      onChanged: (val){
+        setState(() => {
+          if (label == 'First Name') fname = val,
+          if (label == 'Last Name') lname = val,
+          if (label == 'Username') username = val,
+          if (label == 'E-mail') email = val,
+          if (label == 'Password') password = val,
+          if (label == 'University') uni = val
+        });
+      },
       obscureText: obscureText,
       decoration: InputDecoration(
         labelText: label,
-        border: const OutlineInputBorder(),
+        labelStyle: const TextStyle(
+          fontSize: 18,
+          color: Colors.black,
+        ),
+        border: InputBorder.none,
+        fillColor: Colors.black12,
+        filled: true,
       ),
     );
+  }
+
+  String? validatePassword(String? value) {
+    if (value == null || value.isEmpty) {
+      return "Enter a password";
+    } else if (value.length < 8) {
+      return "Password must be at least 8 characters long";
+    } else if (!RegExp(r'(?=.*?[A-Z])').hasMatch(value)) {
+      return "Password must have at least one uppercase letter";
+    } else if (!RegExp(r'(?=.*?[a-z])').hasMatch(value)) {
+      return "Password must have at least one lowercase letter";
+    } else if (!RegExp(r'(?=.*?[0-9])').hasMatch(value)) {
+      return "Password must have at least one number";
+    } else if (!RegExp(r'(?=.*?[!@#\$&*~]).{6,}').hasMatch(value)) {
+      return "Password must have at least one special character";
+    }
+    return null;
   }
 }
